Consolidate related tests into fewer files
