+++ teste.pas
@@ -1,0 +1,7 @@
+namespace Test;
+
+type
+  Dummy = class
+  end;
+
+end.
