--- tests/AsCast.cs
@@ -1,9 +1,0 @@
-namespace Demo {
-    public partial class AsExample {
-        public string CastIt(object v) {
-            string s;
-            s = v as string;
-            return s;
-        }
-    }
-}
