--- tests/CharCodeSeq.cs
@@ -1,7 +1,0 @@
-namespace Demo {
-    public partial class CharSeq {
-        public static string GetCRLF() {
-            return "\r\n";
-        }
-    }
-}
