+++ tests/MiscOps.pas
@@ -10,6 +10,11 @@ type
     class method Run(ds: DataSet);
   end;
 
+  RoundDouble = public class
+  public
+    class method Adjust(qt_quebra: Integer);
+  end;
+
 implementation
 
 class method CastExample.Run(ds: DataSet);
@@ -21,6 +26,12 @@ begin
     log.Error('QUERY STRING:' + sql);
     log.Error('LINHAS DA QUERY: "' + ds.Tables[0].Rows.Count.ToString + '"');
   end;
+end;
+
+class method RoundDouble.Adjust(qt_quebra: Integer);
+begin
+  if qt_quebra > 10 then
+    qt_quebra := Round(double(qt_quebra / 3))-1;
 end;
 
 end.
