--- tests/NamedArg.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Demo {
-    public partial class NamedArg {
-        public void UseArgs() {
-            Console.WriteLine(123, "ok");
-        }
-    }
-}
