--- tests/VarInfer.cs
@@ -1,7 +1,0 @@
-namespace Demo {
-    public partial class VarInfer {
-        public void Example() {
-            var cli = new WebClient();
-        }
-    }
-}
