--- tests/VarStmt.cs
@@ -1,7 +1,0 @@
-namespace Demo {
-    public partial class VarStmt {
-        public void Example() {
-            string delimiter = ",";
-        }
-    }
-}
