--- tests/VarStmt.pas
@@ -1,16 +1,0 @@
-namespace Demo;
-
-type
-  VarStmt = public class
-  public
-    method Example;
-  end;
-
-implementation
-
-method VarStmt.Example;
-begin
-  var delimiter : String := ',';
-end;
-
-end.
